clean up - broke out larger components into own file - created interfaces for components
